Fix msvc warnings

+++ modules/robot/src/real-robot/afma4/vpRobotAfma4.cpp
@@ -1650,7 +1650,7 @@ vpRobotAfma4::readPosFile(const std::string &filename, vpColVector &q)
       std::istringstream ss(line);
       std::string key_;
       ss >> key_;
-      for (size_t i=0; i< njoint; i++)
+      for (unsigned int i=0; i< njoint; i++)
         ss >> q[i];
       pos_found = true;
       break;
