+++ modules/robot/src/real-robot/ptu46/vpRobotPtu46.cpp
@@ -806,7 +806,7 @@ vpRobotPtu46::readPositionFile(const std::string &filename, vpColVector &q)
       std::istringstream ss(line);
       std::string key_;
       ss >> key_;
-      for (size_t i=0; i< vpPtu46::ndof; i++)
+      for (unsigned int i=0; i< vpPtu46::ndof; i++)
         ss >> q[i];
       pos_found = true;
       break;
