+++ modules/robot/src/robot-simulator/vpSimulatorViper850.cpp
@@ -2095,7 +2095,7 @@ vpSimulatorViper850::readPosFile(const std::string &filename, vpColVector &q)
       std::istringstream ss(line);
       std::string key_;
       ss >> key_;
-      for (size_t i=0; i< njoint; i++)
+      for (unsigned int i=0; i< njoint; i++)
         ss >> q[i];
       pos_found = true;
       break;
